refactor: 버튼 분리
