+++ package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "axios": "^1.6.8",
+    "date-fns": "^3.6.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-query": "^3.39.3",
