+++ src/pages/dashboard/DashboardPage.tsx
@@ -3,14 +3,24 @@ import TopNavigation from '@layout/TopNavigation';
 import BottomNavigation from '@layout/BottomNavigation';
 import type { NavLayoutProps } from '../../types/navigationTypes';
 
+import { useNavigate } from 'react-router-dom';
+
 const NavigationLayout = ({ children }: NavLayoutProps) => {
+  const navigate = useNavigate();
+
   return (
     <>
       <TopNavigation
         _TopBar={
           <TopNavigation.TopBar
             centerContent={<div>대시보드</div>}
-            rightContent={TopNavigation.TopBar.SettingButton}
+            rightContent={
+              <TopNavigation.TopBar.SettingButton
+                onClick={() => {
+                  navigate('/setting');
+                }}
+              />
+            }
           />
         }></TopNavigation>
       {children}
